+++ src/components/Calendar.jsx
@@ -1,138 +1,66 @@
-import { useState, useEffect } from "react";
-import {
-  endOfWeek,
-  startOfWeek,
-  isWithinInterval,
-  addDays,
-  eachDayOfInterval,
-  formatISO9075,
-  format,
-} from "date-fns";
 import { DayPicker } from "react-day-picker";
-import "react-day-picker/src/style.css";
-import DaysOff from "./DaysOff";
+import { useEffect } from "react";
 
-// used for producing dates in ISO format to be used as id
-const cleanUpDates = (days) => {
-  const cleanedUpDays = [];
-  for (let i = 0; i < 7; i += 1) {
-    cleanedUpDays.push(formatISO9075(days[i], { representation: "date" }));
-  }
-  return cleanedUpDays;
-};
+import DaysOff from "./DaysOff";
+import useCalendar from "../hooks/useCalendar";
+
+import "react-day-picker/src/style.css";
 
 export default function Calendar({ onSelectedDaysChange, onOffDaysChange }) {
-  // defaults highlighted week to next week
-  const [selectedDays, setSelectedDays] = useState({
-    from: startOfWeek(addDays(new Date(), 7), { weekStartsOn: 1 }),
-    to: endOfWeek(addDays(new Date(), 7), { weekStartsOn: 1 }),
-  });
+    const {
+        handleDayClick,
+        handleDaysOffChange,
+        formatWeekdayName,
+        selectedDays,
+        isWithinInterval,
+        offDaysDisabled,
+        selectedDaysData,
+        offDays,
+    } = useCalendar();
 
-  // produces array of dates in ISO format to be used as id
-  const [selectedDaysData, setSelectedDaysData] = useState(
-    cleanUpDates(
-      eachDayOfInterval({
-        start: startOfWeek(addDays(new Date(), 7), { weekStartsOn: 1 }),
-        end: endOfWeek(addDays(new Date(), 7), { weekStartsOn: 1 }),
-      })
-    )
-  );
+    // console logs dates when week is clicked
+    useEffect(() => {
+        console.log(selectedDaysData);
+        console.log(offDays);
+    }, [selectedDaysData, offDays]);
 
-  const handleDayClick = (day) => {
-    if (
-      selectedDays &&
-      isWithinInterval(day, { start: selectedDays.from, end: selectedDays.to })
-    ) {
-      setSelectedDays(null); // Clear selection if clicking within the selected week
-    } else {
-      // used by modifier
-      setSelectedDays({
-        from: startOfWeek(day, { weekStartsOn: 1 }),
-        to: endOfWeek(day, { weekStartsOn: 1 }),
-      });
-      // used for data
-      setSelectedDaysData(
-        cleanUpDates(
-          eachDayOfInterval({
-            start: startOfWeek(day, { weekStartsOn: 1 }),
-            end: endOfWeek(day, { weekStartsOn: 1 }),
-          })
-        )
-      );
-    }
-  };
+    // raises selectedDaysData, offDays to Generate Menu
+    useEffect(() => {
+        onSelectedDaysChange(selectedDaysData);
+        onOffDaysChange(offDays);
+    }, [selectedDaysData, offDays]);
 
-  const formatWeekdayName = (day) => format(day, "ccc");
+    return (
+        <div className="flex flex-col items-center">
+            <DayPicker
+                numberOfMonths={2}
+                weekStartsOn={1}
+                captionLayout={"dropdown"}
+                selected={selectedDays}
+                formatters={{ formatWeekdayName }}
+                modifiers={{
+                    range_start: selectedDays?.from,
+                    range_end: selectedDays?.to,
+                    range_middle: (date) =>
+                        selectedDays
+                            ? isWithinInterval(date, {
+                                  start: selectedDays.from,
+                                  end: selectedDays.to,
+                              })
+                            : false,
+                }}
+                disabled={[{ dayOfWeek: offDaysDisabled }, { before: new Date() }]}
+                onDayClick={handleDayClick}
+            />
+            {selectedDays && (
+                <p>
+                    Week from {selectedDays.from.toLocaleDateString()} to {selectedDays.to.toLocaleDateString()}
+                </p>
+            )}
+            <DaysOff onDaysOffChange={handleDaysOffChange} />
 
-  // default off days
-  const [offDays, setOffDays] = useState([
-    false,
-    false,
-    false,
-    false,
-    false,
-    true,
-    true,
-  ]);
-  const [offDaysDisabled, setOffDaysDisabled] = useState([6, 0]);
-
-  const handleDaysOffChange = (offDaysArray) => {
-    setOffDays(offDaysArray);
-  };
-
-  // updates the disabled weekdays
-  useEffect(() => {
-    setOffDaysDisabled(
-      offDays.flatMap((x, index) =>
-        x ? (index + 1 === 7 ? 0 : index + 1) : []
-      )
-    );
-  }, [offDays]);
-
-  // console logs dates when week is clicked
-  useEffect(() => {
-    console.log(selectedDaysData);
-    console.log(offDays);
-  }, [selectedDaysData, offDays]);
-
-  // raises selectedDaysData, offDays to Generate Menu
-  useEffect(() => {
-    onSelectedDaysChange(selectedDaysData);
-    onOffDaysChange(offDays);
-  }, [selectedDaysData, offDays]);
-
-  return (
-    <div className="flex flex-col items-center">
-      <DayPicker
-        numberOfMonths={2}
-        weekStartsOn={1}
-        captionLayout={"dropdown"}
-        selected={selectedDays}
-        formatters={{ formatWeekdayName }}
-        modifiers={{
-          range_start: selectedDays?.from,
-          range_end: selectedDays?.to,
-          range_middle: (date) =>
-            selectedDays
-              ? isWithinInterval(date, {
-                  start: selectedDays.from,
-                  end: selectedDays.to,
-                })
-              : false,
-        }}
-        disabled={[{ dayOfWeek: offDaysDisabled }, { before: new Date() }]}
-        onDayClick={handleDayClick}
-      />
-      {selectedDays && (
-        <p>
-          Week from {selectedDays.from.toLocaleDateString()} to{" "}
-          {selectedDays.to.toLocaleDateString()}
-        </p>
-      )}
-      <DaysOff onDaysOffChange={handleDaysOffChange} />
-
-      <style>
-        {`
+            <style>
+                {`
             .rdp-root {
                 --rdp-accent-color: #DDEEF8;
                 --rdp-outside-opacity: 0.35; 
@@ -157,7 +85,7 @@ export default function Calendar({ onSelectedDaysChange, onOffDaysChange }) {
             }
           }
         `}
-      </style>
-    </div>
-  );
+            </style>
+        </div>
+    );
 }
